+++ cypress/integration/EventWorkingLists/EventWorkingListsDev/index.js
@@ -154,7 +154,7 @@ Then('active events that are assigned to anyone should be retrieved from the api
 
 When('you set the age filter to 10-20', () => {
     cy.get('[data-test="event-working-lists"]')
-        .contains('Age (years)')
+        .contains('Age in years')
         .click();
 
     cy.get('[data-test="list-view-filter-contents"]')
@@ -175,7 +175,7 @@ When('you apply the current filter', () => {
 
 Then('the age filter button should show that the filter is in effect', () => {
     cy.get('[data-test="event-working-lists"]')
-        .contains('Age (years): 10 to 20')
+        .contains('Age in years: 10 to 20')
         .should('exist');
 });