+++ cypress/integration/NewPage/index.js
@@ -36,10 +36,7 @@ And('you select tracked entity type person', () => {
 
 And('you see the registration form for the Person', () => {
     cy.get('[data-test="dhis2-capture-registration-page-content"]')
-        .contains('New Person')
-        .should('exist');
-    cy.get('[data-test="dhis2-capture-registration-page-content"]')
-        .contains('New Person')
+        .contains('New person')
         .should('exist');
     cy.get('[data-test="dhis2-capture-registration-page-content"]')
         .contains('Profile')
@@ -119,7 +116,7 @@ And('you see the registration form for the Malaria case registration', () => {
         .contains('Coordinate')
         .should('exist');
     cy.get('[data-test="dhis2-capture-registration-page-content"]')
-        .contains('Age (years)')
+        .contains('Age in years')
         .should('exist');
     cy.get('[data-test="dhis2-capture-registration-page-content"]')
         .contains('Household location')