+++ src/core_modules/capture-core/components/LockedSelector/QuickSelector/QuickSelector.component.js
@@ -2,20 +2,30 @@
 
 import React, { Component } from 'react';
 import { withStyles } from '@material-ui/core/styles';
-
+import i18n from '@dhis2/d2-i18n';
 import Paper from '@material-ui/core/Paper';
 import Grid from '@material-ui/core/Grid';
-
+import { colors } from '@dhis2/ui';
 import programs from 'capture-core/metaDataMemoryStores/programCollection/programCollection';
 
 import ProgramSelector from './Program/ProgramSelector.component';
 import OrgUnitSelector from './OrgUnitSelector.component';
 import { ActionButtons } from './ActionButtons.component';
+import { SingleLockedSelect } from './SingleLockedSelect.component';
 
-const styles = () => ({
+const styles = ({ palette }) => ({
     paper: {
         flexGrow: 1,
-        padding: 10,
+    },
+    programSelector: {
+        backgroundColor: palette.grey.lighter,
+        margin: '0 0 0 -1px',
+    },
+    orgUnitSelector: {
+        backgroundColor: palette.grey.lighter,
+        margin: '0 0 0 -1px',
+        borderRight: `1px solid ${colors.grey500}`,
+        borderLeft: `1px solid ${colors.grey500}`,
     },
 });
 
@@ -37,6 +47,15 @@ type Props = {
     onNewClickWithoutProgramId: () => void,
     onFindClick: () => void,
     onFindClickWithoutProgramId: () => void,
+    currentPage: string,
+    selectedEnrollmentId: string,
+    selectedTeiName: string,
+    selectedTetName: string,
+    enrollmentsAsOptions: Array<Object>,
+    onTeiSelectionReset: () => void,
+    onEnrollmentSelectionSet: () => void,
+    onEnrollmentSelectionReset: () => void,
+    enrollmentLockedSelectReady: boolean,
 };
 
 class QuickSelector extends Component<Props> {
@@ -72,38 +91,30 @@ class QuickSelector extends Component<Props> {
         const selectedProgramId = this.props.selectedProgramId;
         const selectedProgram = QuickSelector.getSelectedProgram(selectedProgramId);
 
-        let orgUnitSelectorWidth = 3;
-        let programSelectorWidth = 3;
-        let actionButtonsWidth = 6;
-
-        if (selectedProgram && selectedProgram.categoryCombination) {
-            orgUnitSelectorWidth = 3;
-            programSelectorWidth = 6;
-            actionButtonsWidth = 3;
-        }
-
         return {
-            orgUnitSelectorWidth,
-            programSelectorWidth,
-            actionButtonsWidth,
+            programSelectorWidth: selectedProgram && selectedProgram.categoryCombination ? 4 : 2,
+            width: 2,
         };
     }
 
     render() {
-        const { orgUnitSelectorWidth, programSelectorWidth, actionButtonsWidth } = this.calculateColumnWidths();
+        const { width, programSelectorWidth } = this.calculateColumnWidths();
+        const {
+            currentPage,
+            selectedTeiName,
+            selectedTetName,
+            enrollmentsAsOptions,
+            selectedEnrollmentId,
+            onTeiSelectionReset,
+            onEnrollmentSelectionSet,
+            onEnrollmentSelectionReset,
+            enrollmentLockedSelectReady,
+        } = this.props;
 
         return (
             <Paper className={this.props.classes.paper}>
-                <Grid container spacing={16}>
-                    <Grid item xs={12} sm={orgUnitSelectorWidth}>
-                        <OrgUnitSelector
-                            selectedOrgUnitId={this.props.selectedOrgUnitId}
-                            handleClickOrgUnit={this.handleClickOrgUnit}
-                            selectedOrgUnit={this.props.selectedOrgUnit}
-                            onReset={this.props.onResetOrgUnitId}
-                        />
-                    </Grid>
-                    <Grid item xs={12} sm={programSelectorWidth}>
+                <Grid container spacing={0}>
+                    <Grid item xs={12} sm={programSelectorWidth * 3} md={programSelectorWidth * 2} lg={programSelectorWidth} className={this.props.classes.programSelector}>
                         <ProgramSelector
                             selectedProgram={this.props.selectedProgramId}
                             selectedOrgUnitId={this.props.selectedOrgUnitId}
@@ -117,7 +128,47 @@ class QuickSelector extends Component<Props> {
                             onResetOrgUnit={this.props.onResetOrgUnitId}
                         />
                     </Grid>
-                    <Grid item xs={12} sm={actionButtonsWidth}>
+                    <Grid item xs={12} sm={width * 3} md={width * 2} lg={width} className={this.props.classes.orgUnitSelector}>
+                        <OrgUnitSelector
+                            selectedOrgUnitId={this.props.selectedOrgUnitId}
+                            handleClickOrgUnit={this.handleClickOrgUnit}
+                            selectedOrgUnit={this.props.selectedOrgUnit}
+                            onReset={this.props.onResetOrgUnitId}
+                        />
+                    </Grid>
+                    {
+                        currentPage === 'enrollment' &&
+                        <>
+                            <Grid item xs={12} sm={width * 3} md={width * 2} lg={2} className={this.props.classes.orgUnitSelector}>
+                                <SingleLockedSelect
+                                    ready={enrollmentLockedSelectReady}
+                                    onClear={onTeiSelectionReset}
+                                    options={[{
+                                        label: selectedTeiName,
+                                        value: 'alwaysPreselected',
+
+                                    }]}
+                                    selectedValue="alwaysPreselected"
+                                    title={selectedTetName}
+                                />
+                            </Grid>
+                            {
+                                enrollmentsAsOptions &&
+                                <Grid item xs={12} sm={width * 3} md={width * 2} lg={2} className={this.props.classes.orgUnitSelector}>
+                                    <SingleLockedSelect
+                                        onClear={onEnrollmentSelectionReset}
+                                        ready={enrollmentLockedSelectReady}
+                                        onSelect={onEnrollmentSelectionSet}
+                                        options={enrollmentsAsOptions}
+                                        selectedValue={selectedEnrollmentId}
+                                        title={i18n.t('Enrollment')}
+                                    />
+                                </Grid>
+                            }
+                        </>
+
+                    }
+                    <Grid item xs={12} sm={width * 3} md={width * 2} lg={2} >
                         <ActionButtons
                             selectedProgramId={this.props.selectedProgramId}
                             onStartAgainClick={this.props.onStartAgain}