+++ src/core_modules/capture-core/components/Pages/Enrollment/EnrollmentPage.types.js
@@ -1,0 +1,14 @@
+// @flow
+import { typeof enrollmentPageStatuses } from './EnrollmentPage.constants';
+
+export type EnrollmentPageStatus = $Keys<enrollmentPageStatuses>
+
+export type ContainerProps = $ReadOnly<{|
+  error: boolean,
+  enrollmentPageStatus: EnrollmentPageStatus,
+|}>
+
+export type Props = {|
+  ...ContainerProps,
+  ...CssClasses
+|}