+++ src/core_modules/capture-core/components/Pages/MainPage/EventWorkingLists/TemplateSetup/EventWorkingListsTemplateSetup.component.js
@@ -39,11 +39,10 @@ export const EventWorkingListsTemplateSetup = ({
         });
     }, [onUpdateTemplate, filters, columns, sortById, sortByDirection, program.id]);
 
-    const injectArgumentsForAddTemplate = React.useCallback((name, template) => {
+    const injectArgumentsForAddTemplate = React.useCallback((name) => {
         // $FlowFixMe For columns: fixing this will create really ugly code. SortById, sortByDirection: Rather complex logic results in sortById and sortByDirection always having a value here.
         const eventQueryCriteria = convertToEventFilterEventQueryCriteria({ filters, columns: new Map(columns.map(c => [c.id, c])), sortById, sortByDirection });
         onAddTemplate(name, eventQueryCriteria, {
-            template,
             filters,
             visibleColumnIds: columns && columns
                 .filter(({ visible }) => visible)