+++ src/core_modules/capture-core/components/Pages/MainPage/WorkingLists/TemplateMaintenance/sharingDialog.types.js
@@ -1,0 +1,9 @@
+// @flow
+import type { SharingSettings } from '../workingLists.types';
+
+export type Props = {|
+    onClose: (sharingSettings: SharingSettings) => void,
+    open: boolean,
+    templateId: string,
+    ...CssClasses,
+|};