+++ src/core_modules/capture-core/components/Pages/MainPage/WorkingLists/TemplateMaintenance/templateMaintenance.types.js
@@ -1,0 +1,23 @@
+// @flow
+import { typeof dialogModes } from './dialogModes';
+import type {
+    WorkingListTemplate,
+    SetTemplateSharingSettings,
+    AddTemplate,
+    UpdateTemplate,
+    DeleteTemplate,
+} from '../workingLists.types';
+
+export type PassOnProps = {
+    onClose: () => void,
+};
+
+export type Props = {
+    ...PassOnProps,
+    mode: ?$Values<dialogModes>,
+    currentTemplate: WorkingListTemplate,
+    onAddTemplate: AddTemplate,
+    onUpdateTemplate: UpdateTemplate,
+    onDeleteTemplate: DeleteTemplate,
+    onSetSharingSettings: SetTemplateSharingSettings,
+};