+++ src/core_modules/capture-core/metaDataMemoryStoreBuilders/programs/factory/programStage/ProgramStageFactory.js
@@ -127,6 +127,11 @@ class ProgramStageFactory {
                 cachedProgramStage.id,
             );
             _stage.enableUserAssignment = !!cachedProgramStage.enableUserAssignment;
+            _stage.autoGenerateEvent = !!cachedProgramStage.autoGenerateEvent;
+            _stage.openAfterEnrollment = !!cachedProgramStage.openAfterEnrollment;
+            _stage.generatedByEnrollmentDate = !!cachedProgramStage.generatedByEnrollmentDate;
+            _stage.reportDateToUse = cachedProgramStage.reportDateToUse;
+            _stage.standardInterval = cachedProgramStage.standardInterval;
             _stage.stageForm = new RenderFoundation((_form) => {
                 _form.id = cachedProgramStage.id;
                 _form.name = cachedProgramStage.displayName;